Hbv4 Stream tests addition (#12)
* modify install script and add hbv4 conf

* more modifications to entire directory

* work on stream

* work on stream

* stream build workflow

* saving work

* modifications for stream

* fix bug I introduced last PR

---------

+++ customTests/custom-test-setup.sh
@@ -17,40 +17,63 @@ fi
 
 mkdir -p $EXE_DIR
 
-# CUDA BW Test Setup
-#Test if nvcc is installed and if so install gpu-copy test.
-if test -f "$NVCC"; then
-	#Compile the gpu-copy benchmark.
+#Nvidia installs
+if [ -e '/dev/nvidiactl' ]; then
+	# CUDA BW Test Setup
+	#Test if nvcc is installed and if so install gpu-copy test.
+	if test -f "$NVCC"; then
+		#Compile the gpu-copy benchmark.
 
-	cufile="$SRC_DIR/gpu-copy.cu"
-	outfile="$EXE_DIR/gpu-copy"
+		cufile="$SRC_DIR/gpu-copy.cu"
+		outfile="$EXE_DIR/gpu-copy"
 
-	#Test if the default gcc compiler is new enough to compile gpu-copy.
-	#If it is not then use the 9.2 compiler, that should be installed in
-	#/opt.
-	if [ $(gcc -dumpversion | cut -d. -f1) -gt 6 ]; then
-		$NVCC -lnuma $cufile -o $outfile
+		#Test if the default gcc compiler is new enough to compile gpu-copy.
+		#If it is not then use the 9.2 compiler, that should be installed in
+		#/opt.
+		if [ $(gcc -dumpversion | cut -d. -f1) -gt 6 ]; then
+			$NVCC -lnuma $cufile -o $outfile
+		else
+			$NVCC --compiler-bindir /opt/gcc-9.2.0/bin \
+				-lnuma $cufile -o $outfile
+		fi
 	else
-		$NVCC --compiler-bindir /opt/gcc-9.2.0/bin \
-			-lnuma $cufile -o $outfile
+  		echo "$NVCC not found. Exiting setup"
 	fi
-else
-  echo "$NVCC not found. Exiting setUP"
+
+	# create perf-test executables
+	echo -e "Building PerfTest"
+	VERSION=4.5-0.12
+	VERSION_HASH=ge93c538
+	apt-get install -y libpci-dev
+	pushd ${EXE_DIR}
+	wget https://github.com/linux-rdma/perftest/releases/download/v${VERSION}/perftest-${VERSION}.${VERSION_HASH}.tar.gz
+	tar xvf perftest-${VERSION}.${VERSION_HASH}.tar.gz
+	pushd perftest-4.5
+	./configure CUDA_H_PATH=/usr/local/cuda/include/cuda.h
+	make
+	rm ${EXE_DIR}/perftest-${VERSION}.${VERSION_HASH}.tar.gz
+	popd
+	popd
 fi
 
-# create perf-test executables
-VERSION=4.5-0.12
-VERSION_HASH=ge93c538
-apt-get install -y libpci-dev
-pushd ${EXE_DIR}
-wget https://github.com/linux-rdma/perftest/releases/download/v${VERSION}/perftest-${VERSION}.${VERSION_HASH}.tar.gz
-tar xvf perftest-${VERSION}.${VERSION_HASH}.tar.gz
-pushd perftest-4.5
-./configure CUDA_H_PATH=/usr/local/cuda/include/cuda.h
-make
-rm ${EXE_DIR}/perftest-${VERSION}.${VERSION_HASH}.tar.gz
-popd
-popd
+# Stream
+if command -v /opt/AMD/aocc-compiler-4.0.0/bin/clang &> /dev/null || command -v clang &> /dev/null; then
+	echo -e "clang compiler found Building Stream"
+	pushd ${SRC_DIR}/stream
+	if ! [[ -f "stream.c" ]]; then 
+		wget https://www.cs.virginia.edu/stream/FTP/Code/stream.c
+	fi
+
+	if command -v /opt/AMD/aocc-compiler-4.0.0/bin/clang &> /dev/null; then
+		make CC=/opt/AMD/aocc-compiler-4.0.0/bin/clang EXEC_DIR=$EXE_DIR
+	else
+		make CC=clang EXEC_DIR=$EXE_DIR
+	fi
+	popd
+else
+  echo "clang command not found"
+fi
+
 # copy all custom test to the nhc scripts dir
 cp $SRC_DIR/*.nhc /etc/nhc/scripts
